pyproject.toml
[tool.poetry]
name = "flamethrower"
packages = [{ from = "src", include = "flamethrower" }]
version = "0.2.13"
description = "The ultimate debugging experience 🦠🔥"
authors = ["Scott Susanto <scottsus@usc.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
flamethrower = "flamethrower.shell:shell.run"
ft = "flamethrower.shell:shell.run"

[tool.poetry.dependencies]
python = "^3.08"
pydantic = "^2.5.3"
pathspec = "^0.12.1"
openai = "^1.6.1"
jsonschema = "^4.20.0"
rich = "^13.7.0"
tiktoken = "^0.5.2"
gitpython = "^3.1.40"
backoff = "^2.2.1"
questionary = "^2.0.1"
dependency-injector = "^4.41.0"
importlib-resources = "^6.1.1"
types-jsonschema = "^4.21.0.20240118"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pytest = "^8.0.0"
pyautogui = "^0.9.54"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
